a bunch of refactoring
