+++ pyxl/codec/register.py
@@ -1,44 +1,10 @@
 #!/usr/bin/env python
 
 
-import codecs, io, encodings
+import codecs, encodings
 import sys
-import traceback
-from encodings import utf_8
-from pyxl.codec.tokenizer import pyxl_tokenize, pyxl_untokenize
+from pyxl.codec.transform import pyxl_decode, PyxlIncrementalDecoder, PyxlStreamReader
 
-def pyxl_transform(stream):
-    try:
-        output = pyxl_untokenize(pyxl_tokenize(stream.readline))
-    except Exception as ex:
-        print(ex)
-        traceback.print_exc()
-        raise
-
-    return output.rstrip()
-
-def pyxl_transform_string(input):
-    stream = io.StringIO(bytes(input).decode('utf-8'))
-    return pyxl_transform(stream)
-
-def pyxl_decode(input, errors='strict'):
-    return pyxl_transform_string(input), len(input)
-
-class PyxlIncrementalDecoder(utf_8.IncrementalDecoder):
-    def decode(self, input, final=False):
-        self.buffer += input
-        if final:
-            buff = self.buffer
-            self.buffer = b''
-            return super(PyxlIncrementalDecoder, self).decode(
-                pyxl_transform_string(buff).encode('utf-8'), final=True)
-        else:
-            return ''
-
-class PyxlStreamReader(utf_8.StreamReader):
-    def __init__(self, *args, **kwargs):
-        codecs.StreamReader.__init__(self, *args, **kwargs)
-        self.stream = io.StringIO(pyxl_transform(self.stream))
 
 def search_function(encoding):
     if encoding != 'pyxl': return None
