+++ pyxl/codec/register_invertible.py
@@ -5,76 +5,11 @@ import codecs, io, encodings
 import sys
 import traceback
 from encodings import utf_8
-from pyxl.codec.tokenizer import (
-    pyxl_reverse_tokenize, pyxl_tokenize, pyxl_untokenize,
-    PyxlUnfinished,
+from pyxl.codec.transform import (
+    pyxl_encode, pyxl_decode, PyxlIncrementalDecoderInvertible, PyxlIncrementalEncoder,
+    PyxlStreamReaderInvertible, PyxlStreamWriter,
 )
 
-def pyxl_transform(stream):
-    try:
-        output = pyxl_untokenize(pyxl_tokenize(stream.readline, invertible=True))
-    except Exception as ex:
-        print(ex)
-        traceback.print_exc()
-        raise
-
-    return output
-
-def pyxl_reverse(stream):
-    try:
-        output = pyxl_untokenize(pyxl_reverse_tokenize(stream.readline))
-    except PyxlUnfinished:
-        raise
-    except Exception as ex:
-        print(ex)
-        traceback.print_exc()
-        raise
-
-    return output.encode('utf-8')
-
-def pyxl_transform_string(input):
-    stream = io.StringIO(bytes(input).decode('utf-8'))
-    return pyxl_transform(stream)
-
-def pyxl_reverse_string(input):
-    stream = io.StringIO(input)
-    return pyxl_reverse(stream)
-
-def pyxl_encode(input, errors='strict'):
-    # FIXME: maybe we should actually be able to consume partial results
-    # instead of this O(n^2) retry thing?
-    try:
-        return pyxl_reverse_string(input), len(input)
-    except PyxlUnfinished:
-        return b'', 0
-
-def pyxl_decode(input, errors='strict'):
-    return pyxl_transform_string(input), len(input)
-
-class PyxlIncrementalDecoder(utf_8.IncrementalDecoder):
-    def decode(self, input, final=False):
-        self.buffer += input
-        if final:
-            buff = self.buffer
-            self.buffer = b''
-            return super(PyxlIncrementalDecoder, self).decode(
-                pyxl_transform_string(buff).encode('utf-8'), final=True)
-        else:
-            return ''
-
-class PyxlIncrementalEncoder(codecs.BufferedIncrementalEncoder):
-    def _buffer_encode(self, input, errors, final):
-        return pyxl_encode(input, errors)
-
-
-class PyxlStreamReader(utf_8.StreamReader):
-    def __init__(self, *args, **kwargs):
-        codecs.StreamReader.__init__(self, *args, **kwargs)
-        self.stream = io.StringIO(pyxl_transform(self.stream))
-
-
-class PyxlStreamWriter(codecs.StreamWriter):
-    encode = pyxl_encode
 
 def search_function(encoding):
     if encoding != 'pyxl': return None
@@ -83,10 +18,10 @@ def search_function(encoding):
     return codecs.CodecInfo(
         name = 'pyxl',
         encode = pyxl_encode,
-        decode = pyxl_decode,
+        decode = lambda b: pyxl_decode(b, invertible=True),
         incrementalencoder = PyxlIncrementalEncoder,
-        incrementaldecoder = PyxlIncrementalDecoder,
-        streamreader = PyxlStreamReader,
+        incrementaldecoder = PyxlIncrementalDecoderInvertible,
+        streamreader = PyxlStreamReaderInvertible,
         streamwriter = PyxlStreamWriter,
     )
