+++ tests/test_roundtrips.py
@@ -1,4 +1,4 @@
-from pyxl.codec.register_invertible import pyxl_encode, pyxl_decode
+from pyxl.codec.transform import pyxl_transform_string, pyxl_reverse_string
 
 import os
 
@@ -6,9 +6,10 @@ dir_path = os.path.dirname(os.path.abspath(__file__))
 
 def _roundtrip(file_name):
     path = os.path.join(dir_path, file_name)
-    with open(path, "rb") as f:
+    with open(path, "r") as f:
         contents = f.read()
-        assert contents == pyxl_encode(pyxl_decode(contents)[0])[0], (
+        depyxled = pyxl_transform_string(contents, invertible=True)
+        assert contents == pyxl_reverse_string(depyxled), (
             "Could not round-trip file %s" % file_name)
 
 # TODO: it would be better if each file was automatically a separate test case...
